Moved pwd checks into a subfolder
