+++ frontend/src/components/utilities/checks/password/checkPassword.ts
