+++ frontend/src/components/utilities/checks/password/passwordRegexes.ts
